Handles new installer arguments and creates friendly organization keys.
The new Huntress Agent Installer requires and account key and an organization key. This update handles the new arguments. Based on partner feedback, we've created a hybrid way to generate organization keys based on Continuum's SITEID and SITENAME values. See the comment within the deployment script (below the legal disclaimer) for full details.