Service check improvements, EDR/kernel 6.2+ check
Added kernel 6.2 checks so that when we check for running and installed services we check for EDR as well. Inform the user and log if EDR will not be available.
Adjusted install timeout value to 120 and moved it to the "user modified variables" section.
Improved AccountKey and OrgKey field checks, should no longer be possible to install with the placeholder value.
Squashed a couple bugs that lead to incomplete logs.