Improved self-repair
When Huntress is uninstalled or damaged the services are often left in the "disabled" startup state. Partners attempt to repair this by reinstalling Huntress, but if the services are set to "disabled" start then they can't be manually started. We can address that by setting all non-auto start services to start automatically. SC.exe was used for compatibility, the better PoSh commands aren't available until 6.0 (most people are on 5.0).