Update InstallHuntress.powershellv2.ps1
Hardcoding the length of $AccountKey was causing some issues during uninstall since the script assumed a length of 32. By pivoting to using the actual string length when masking the value for logging we can catch edge cases where an invalid value was passed or if no value was passed such as during uninstall.