additional improvement to account key length checks
Catch $AccountKey length < 8 to avoid errors and display the value of that string in a non-masked fashion to make it easier for Support and partner to identify why the key is invalid. Should be helpful especially for partners using RMM's and other automations which can be difficult to understand what parameters are being passed.