Update InstallHuntress.dattormm.comstore.ps1
Datto script uses TagsKey instead of Tags, so that should be logged just in case there are invalid characters. 

Also moved the commands that log keys further up in the chain to make sure they're captured before Test-Parameters is run, this is needed to ensure we see the keys in the logs before the script gets to a part of the execution that can be stopped by invalid key formats.